Edit model name.
* Edit Air -> AirResponse, Station -> StationResponse because It is only used for getting response.
