--- app/src/main/java/com/mashupgroup/weatherbear/models/air/Station.kt
@@ -1,6 +1,0 @@
-package com.mashupgroup.weatherbear.models.air
-
-data class Station(
-        var totalCount: Int,
-        var list: List<StationItem>
-)
