--- app/src/main/java/com/mashupgroup/weatherbear/models/air/StationItem.kt
@@ -1,7 +1,0 @@
-package com.mashupgroup.weatherbear.models.air
-
-data class StationItem(
-        var stationName: String,
-        var addr: String,
-        var tm: String
-)
