+++ app/src/main/java/com/mashupgroup/weatherbear/models/kakao/Coord.kt
@@ -4,3 +4,12 @@ data class Coord(
         var meta: Meta,
         var documents: List<Documents>
 )
+
+data class Documents(
+        var x: String,
+        var y: String
+)
+
+data class Meta(
+        var total_count: Int
+)
