--- app/src/main/java/com/mashupgroup/weatherbear/models/kakao/Documents.kt
@@ -1,6 +1,0 @@
-package com.mashupgroup.weatherbear.models.kakao
-
-data class Documents(
-        var x: String,
-        var y: String
-)
