--- app/src/main/java/com/mashupgroup/weatherbear/models/kakao/Meta.kt
@@ -1,5 +1,0 @@
-package com.mashupgroup.weatherbear.models.kakao
-
-data class Meta(
-        var total_count: Int
-)
